Don't load belongs_to associations when checking for blank.
Also, don't include linkage data. Both of these changes are needed to avoid N+1 queries, although the latter is due to a bug in JR (for which I have submitted a pull request to fix). But it occurred to me that linkage data isn't that useful if you're not including the relationship anyway.